NN-2334: Move app insights to very beginning of file as per Azure instructions in order to enable traces
