+++ server/azure-appinsights.js
